// src/interfaces/interfaceHistoricoEstoqueEmpresa.ts
export interface historicoEstoqueEmpresaProps {
    hsp_tipo: string;
    pro_nome: string;
    hsp_data: string;
    hsp_antigo: number;
    hsp_movimentado: number;
    hsp_total: number;
}

export interface historicoProdutoEstoqueProps {
    hsp_tipo: string;
    hsp_data: string;
    hsp_antigo: number;
    hsp_movimentado: number;
    hsp_total: number;
}